requirements.txt
dill==0.3.1.1
discord.py==1.6.0
elo==0.1.1
python==3.9.0